my-video-course/tests/integration/requirements.txt
boto3>=1.34.0
pytest>=8.0.0
dnspython>=2.6.0
requests>=2.31.0
